+++ seed/ts-express/nullable/serialization/resources/nullable/service/index.ts
@@ -1,2 +1,3 @@
 export * from "./requests";
 export * as getUsers from "./getUsers";
+export * as deleteUser from "./deleteUser";