+++ seed/ts-express/nullable/serialization/resources/nullable/service/requests/index.ts
@@ -1,1 +1,2 @@
 export { CreateUserRequest } from "./CreateUserRequest";
+export { DeleteUserRequest } from "./DeleteUserRequest";