+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.47.1] - 2025-01-15
+
+- Fix: Resolves an issue where nullable query parameters were not null-safe in their method invocations. The
+  generated code now appropriately guard against `null` values like so:
+
+  ```typescript
+  const _queryParams: Record< ... >;
+  if (value !== undefined) {
+      _queryParams["value"] = value?.toString() ?? null;
+  }
+  ```
+
 ## [0.47.0] - 2025-01-14
 
 - Feature: Add support for `nullable` properties. Users can now specify explicit `null` values